requirements.txt
# Requirements for DDI Tox-Predict Pipeline Phase 1
openai==0.28.0
pandas>=1.5.0
rdkit>=2022.09.1
requests>=2.28.1
python-dotenv>=1.0.0